Fix association_matcher in Rails 4
Rails 4 has deprecated the `:conditions` and `:order` options on
relationships in favor of passing a scope. Those properties are still
supported, but the implementation in Rails 4 shoves them into a scope.
Update the RailsShim class with a method that takes an association
reflection and finds the property to check for `:conditions` and `order`
regardless of rails version.