Loads 'test/unit/assertionfailederror' instead of 'test/unit'.
In ruby 1.8, test/unit.rb adds a Kernel.at_exit hook which is intended to run
the test suite. When using rspec this triggers an empty test suite to be run.
Since the intention of this code is to load Test::Unit::AssertionFailedError,
we can require the file that contains that error class directly.

Also, this removes a ruby 1.9-style hash so things continue to work in ruby
1.8.