Do not require rspec to be required before SM is
Currently, if you want to use shoulda-matchers with RSpec, it must be
declared after RSpec in the Gemfile. This is because SM only integrates
with RSpec if it can see it's available; if RSpec hasn't been loaded yet
then SM will ignore it and move on.

This does not usually cause any issues, because even if SM is declared
before RSpec, RSpec gets loaded before SM anyway. When you run tests,
you use the `rspec` executable, which will require RSpec; and from here
spec_helper will load the remainder of your environment. Therefore, SM
will see that RSpec has been loaded and integrate with it.

However, this does cause issues when you're running your tests with
Spring. Spring works, like any other preloader does, by loading Rails
and the gems in your Gemfile in a sandbox which is cached and kept
separate from the process in which you run your tests. A sharp mind will
notice the problem: if SM is loaded ahead of RSpec, then it will never
integrate with RSpec, even if the `rspec` executable is later used to
run the tests, because once SM is loaded, it's loaded for good.

The only way to solve this is to remove the RSpec detection code which
we have currently, and instead attempt to require RSpec ourselves.