Allow multiple StrongParametersMatchers to exist at once.
The provided RSpec example looks a bit contrived, but it corresponds to this perfectly reasonable Minitest
test case that was not working:

  class UserControllerTest < ActionController::TestCase
    should permit(:name).for(:create)
    should_not permit(:admin).for(:create)
  end

This instantiates two matchers at class load time, which resulted in the second one overriding the double
collection of the first one. Then when the tests are executed, the first matcher would fail, because its
double is not listening to ActionController::Parameters#permit -- only the second matcher's double collection
gets activated.

This is fixed in Doublespeak by only creating one double collection per class.