Remove redundant key value from `create_table`
Before when we called `create_table` within `define_model` we would
delete a key value pair from the `columns` hash. Then we would supply
the same key value pair we deleted to `create_table`. This commit
removes that and makes sure that the deleted key value pair are not
supplied as `options` to `create_table`.