Raise a error when the model is missing some declarations.
When using the `have_many` for example the user maybe have some
misdeclarations on your model, like missed a has many through that dont
declare the has many/belongs to in the through model.
In this case ActiveRecord raise specific excepetions like[1].

With this changes we capture the ActiveRecord exception and re-raise as
a Shoulda::Matchers exception.

[1]:
https://github.com/rails/rails/blob/master/activerecord/lib/active_record/associations.rb#L19