Fix set_flash_matcher now qualifier when is called after the key is set
When the `now` qualifier is called after the key is set, example:

```
expect(controller).not_to set_flash['key for flash'].now
```

Is istantiated a new matcher that does not know about the key that was
set on the `[]`.

This fixes this behavior passing the key explicitly to the matcher.