Fix set_flash_matcher now qualifier when is called after the to is called
When the `now` qualifier is called after the `to` is called, example:

```
set_flash['to for flash.now'].to('value for flash.now').now
```

Is istantiated a new matcher that does not know about the `to` value that was
set before the `now` is called.

This fixes this behavior passing the to explicitly to the matcher.