validate_presence_of works with delegation
If we try define a validation to a attribute defined on a delegated object we
will receive a `NoMethodError` if the delegated attribute expect a collection as
value and not `nil` to a empty value.

Change the definition of what is a collection, before we are checking on
`:has_many` and `:has_and_belongs_to_many` association, now we check only if the
default value of the object is a empty array, since this is the default behavior
of a empty collection on `ActiveRecord`.