numericality: Show msg from only 1st failing submatcher, not all
If multiple submatchers that the numericality matcher creates fail, then
all of their messages will be contained within the message for the
numericality matcher itself. This ends up being too much information,
especially since later, we're going to tack a note on the end of
allow_value's message if we detect that an attribute has changed an
incoming value. If we didn't just show the message for the first failing
submatcher, then that note would be present for all of the submatchers,
and that gets obnoxious fast.