numericality: Do not run all submatchers
This commit changes the numericality matcher so that when running
submatchers, the matcher will stop on the first one that fails instead
of running them all. Since all of the submatchers modify the same
record, running all of them is actually bad since it could possibly make
the failure message confusing (since the record will be in a mixed
state).