Add ValidationMatcherScenario for use in tests
Also add a helper method, `build_scenario_for_validation_matcher`, to
all unit tests.

Our plan is to slowly transition all of the tests to use this helper
method, and that we we can get rid of the umpteen different ways we are
building records across all of the tests.