Remove IneffectiveTestError
Prior to this commit, the numericality matcher raised an
IneffectiveTestError if used against a numeric column (unless other
qualifiers were added). I had a misunderstanding of how the numericality
validation works -- it turns out it doesn't work against the output of
the attribute in question, it works against the input. So raising this
exception is pointless and doesn't actually protect against anything.