Always use doc formatter for RSpec by default
Prior to this commit we only told RSpec to use the doc formatter if we
were running one test file. When a test file is being run using Zeus,
`unit_spec_helper` (and thus `spec_helper`) are loaded ahead of time.
This means that from RSpec's perspective, no files are being run, and so
the documentation formatter is not chosen, which means that the default
progress reporter is used.

This guard was added primarily to ensure that when tests are run on CI,
the progress reporter is used instead of the documentation formatter.
This happens anyway, though, because we pass that manually to RSpec
within the Rakefile. So we don't need this guard.