allow_value: Correctly handle mix of good+bad vals
Prior to 583be38, `allow_value` did not correctly fail when given a
mixture of good and bad values. Assuming it was used in the positive,
when given good values first followed by bad ones, then it failed as
expected, but when given bad values first followed by good ones, then it
passed when it wasn't supposed to.

For example:

``` ruby
class Post
  include ActiveModel::Model
  attr_accessor :color
  validates :color, format: { with: /\A#[a-f0-9]{6}\z/ }
end

describe Post do
  # fails, and should fail
  it { should allow_values('#FFFFFF', '#ffffff').for(:color) }
  # passes, but should fail
  it { should allow_values('#ffffff', '#FFFFFF').for(:color) }
end
```

The aforementioned commit fixed this issue because it made `allow_value`
less stateful than it was. We didn't have tests to cover this behavior,
though, so this commit adds them.