uniqueness: Do not ensure existing record is valid
Currently, the first thing that the uniqueness matcher does is to check
to ensure that the record is valid. This creates a problem because if
the existing record did not exist before and comes from the subject, and
if the model has required attributes which are not set beforehand or
has other validations on the attribute being tested, then the record
will fail validation. Unfortunately, enforcing this will break users'
tests, so roll back this behavior.